Nav Mesh needs testing
I do not know how the package manager's dependencies work.
The files are local on my machine while unity only took in a JSON file.
Trying to drag the folder into unity gives it an error saying that it cannot be part of the assets or library or anywhere in unity really.
I'm am taking the next step in the set-up and I'd like you to fetch the change and check if it works on your machine without the package downloaded to check how package manager works.