UI Manager & Game Manager Updates
UI Manager holds a static instance of itself so that other classes can call RequestUIUpdate(); without needing the entire class to be static or needing direct references to the class itself.

RequestUIUpdate() Takes a string argument of what screen the UI should change it. If the name of that screen is valid then it proceeds, otherwise it throws a debug warning.