Animation test in demo scene
All prefab characters pull from 10 character animators based on the different body types of the character prefabs. At least two of the animators are broken and not given parameters or conditions to transition away from idle.

The first step to generating new enemies will be to determine how its animation situation needs to look so that I know what body types can be created and which can't be done.