removed legacy collider + sword layer order +1
The secondary collider that was previously needed for the tile collision to work has been removed as its job is done by a more elegant solution.

The players weapon sprite order in layer increased from 0 -1 so that the weapon displays over tiles and the player can see what they are hitting.