Simple fixes before build
Options button in UI disabled as it will be not be ready for build.
Game Quit in pause and Exit Game in main menu both close the application.

Player starting location moved from 0,44.5-0 to -3,44.5,0. This is done so that they player can easily see the flag without starting the game covering it up.