Removed legacy damage/healing logic
ThePlayerUIManager previously used a damage/heal structure which caused visual issues with missing health such as:  ❤️❤️💔🤍🤍❤️❤️.
The new structure works as intended without issue and displays missing health as ❤️❤️❤️💔🤍🤍🤍 .

Previous commit had the old method commented out to test back and forth. It is now confirmed that the new method is the way to go and the old method has been removed.